src/main/java/com/userauth/userauth/security/UserAuthUserDetailsService.java
package com.userauth.userauth.security;

import com.userauth.userauth.user.User;
import lombok.AllArgsConstructor;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Component;

import static org.springframework.security.core.userdetails.User.withUsername;

@Component
@AllArgsConstructor
public class UserAuthUserDetailsService implements UserDetailsService {

    private final Users users;
    private final PasswordEncoder passwordEncoder;

    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        User user = users.getUser(username)
                .orElseThrow(() -> new UsernameNotFoundException("Unknown user " + username));

        return withUsername(user.getUsername())
                .password(passwordEncoder.encode(user.getPassword()))
                .authorities(user.getAuthorities())
                .build();
    }
}
